Fallback to IPv4 when IPv6 is unreachable
Multiple users have reported that IPv6 address of rubygems.org is
unreachable[1][2]. We have verified using online tools[3][4] as well
as all of aws regions with ipv6 support that there is no misconfig issue
on our part.
IPv6 route being broken is a bit expected and hence, RFC8305[5]
recommends starting connection to both IPv4 and IPv6 address in parallel
and use whichever returns first (this is TLDR explanation, check RFC for
details). This patch of TCPSocket.initialize does something similar, as
in, it starts connection to all resolved IPs in parallel and uses first
successful connection's IP to initalize the TCPSocket. We use `Net:HTTP`
as our remote featcher and it internally uses TCPSocket.open[6].

Even tho support of IPv6 fallback is spotty, I have open a ruby issue
for it[7].

[1] https://help.rubygems.org/discussions/problems/33671-suggestion-for-easy-solution-for-the-ipv6-related-timeout-problem
[2] https://help.rubygems.org/discussions/problems/31074-timeout-error
[3] https://www.ssllabs.com/ssltest/analyze.html?d=rubygems.org
[4] https://ready.chair6.net/?url=rubygems.org
[5] https://tools.ietf.org/html/rfc8305
[6] https://github.com/ruby/ruby/blob/4444025d16ae1a586eee6a0ac9bdd09e33833f3c/lib/net/http.rb#L950
[7] https://bugs.ruby-lang.org/issues/15628